+++ src/main/java/com/davioooh/bookshelf/repositories/BookRepository.java
@@ -2,9 +2,6 @@ package com.davioooh.bookshelf.repositories;
 
 import com.davioooh.bookshelf.domain.Book;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.data.repository.Repository;
 
-import java.util.List;
-
-public interface BookRepository extends CrudRepository<Book, Integer> {
+public interface BookRepository extends CrudRepository<Book, Long> {
 }
